cython: move shared cdefs into pxd files
To avoid repeated `cdef extern` declarations, move common definitions
into pxd files and then cimport them in the pyx files. From a generated
code/performance perspective, this is a no-op since the cimport
triggers a #include of the relevant C source, just as if the cdef
was present in the pyx file.

This will help with C++ exception annotations by reducing the number of
source locations to modify.